+++ starter-code/src/components/CoolButton.js
@@ -1,39 +1,11 @@
 import React , { Component} from 'react';
-
-const classesList = {
-  isActive: 'is-active',
-  isBlack: 'is-black',
-  isCentered: 'is-centered',
-  isDanger: 'is-danger',
-  isDark: 'is-dark',
-  isFocused: 'is-focused',
-  isGrouped: 'is-grouped',
-  isHovered: 'is-hovered',
-  isInfo: 'is-info',
-  isInverted: 'is-inverted',
-  isLarge: 'is-large',
-  isLight: 'is-light',
-  isLink: 'is-link',
-  isLoading: 'is-loading',
-  isMedium: 'is-medium',
-  isOutlined: 'is-outlined',
-  isPrimary: 'is-primary',
-  isRight: 'is-right',
-  isRounded: 'is-rounded',
-  isSelected: 'is-selected',
-  isSmall: 'is-small',
-  isStatic: 'is-static',
-  isSuccess: 'is-success',
-  isText: 'is-text',
-  isWarning: 'is-warning',
-  isWhite: 'is-white',
-}
+import constants from '../../constants';
 
 export default class CoolButton extends Component {
 
   constructor(props) {
     super(props),
-    this.classes = Object.keys(props).map(prop => classesList[prop]).join(' ')
+    this.classes = Object.keys(props).map(prop => constants.classesList[prop]).join(' ');
   }
 
   render() {
